+++ CHANGELOG.md
@@ -103,6 +103,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#5036](https://github.com/open-telemetry/opentelemetry-python/pull/5036))
 - `opentelemetry-sdk`: Implement experimental Logger configurator
   ([#4980](https://github.com/open-telemetry/opentelemetry-python/pull/4980))
+- `opentelemetry-exporter-otlp-proto-grpc`, `opentelemetry-exporter-otlp-proto-http`: add docker-tests coverage of metrics export
+  ([#5030](https://github.com/open-telemetry/opentelemetry-python/pull/5030))
 
 ## Version 1.40.0/0.61b0 (2026-03-04)